Added Path to Intake 3 balls in 6 ball Autonomous

+++ src/main/java/frc4388/robot/commands/Wait.java
@@ -21,7 +21,7 @@ public class Wait extends CommandBase {
   /**
    * Creates a new WaitCommand.
    */
-  public Wait(SubsystemBase subsystem, float seconds) {
+  public Wait(SubsystemBase subsystem, double seconds) {
     // Use addRequirements() here to declare subsystem dependencies.
 
     m_waitTime = (long) (seconds * 1000);
